feat(shared): organize typings
