+++ packages/shared/constants/Regularization.ts
@@ -1,0 +1,3 @@
+export default {
+  CREATE: 'REGULARIZATION_CREATE'
+}
