+++ packages/shared/typings/scalars.ts
@@ -1,0 +1,1 @@
+export type UUID = string & { readonly _: unique symbol }
